Scene 2
Mouse interaction through as if it were a kinect. The idea is that the
interaction will be performed through a Canvas which will display the UI
(1 render texture and 1 side panel to drag and drop sounds).
the kinect willl allow manipulating a hand object which will project on
the RT of the Canvas

Functions with mouse:
1. Drag and drop sounds
2. flush sounds